refactor: implement PR comments
