+++ src/Nightwatch/rules/min_trade_strength_rule.py
