+++ tests/Nightwatch/test_risk_engine.py
@@ -6,7 +6,7 @@ from Nightwatch.models.risk_decision import RiskDecision
 from Nightwatch.models.risk_engine import RiskEngine
 from Nightwatch.models.signal import Signal
 from Nightwatch.rules.cooldown_rule import CooldownRule
-from Nightwatch.rules.min_trade_strenght_rule import MinTradeStrengthRule
+from Nightwatch.rules.min_trade_strength_rule import MinTradeStrengthRule
 from Nightwatch.rules.risk_rule import RiskRule
 from tests.fixtures.signal_factory import make_signal
